iac/requirements.txt
aws-cdk-lib==2.25.0
constructs>=10.0.0,<11.0.0
aws-lambda-powertools==2.9.0
aws-xray-sdk==2.11.0
pytest==6.2.5
boto3==1.35.97
setuptools==75.1.0
requests==2.32.3
pydantic==2.10.6